requirements.txt
art==6.1
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.9
pydantic==1.10.13
Pygments==2.17.2
python-dotenv==1.0.0
rich==13.7.0
SQLAlchemy==2.0.23
sqlmodel==0.0.12
typer==0.9.0
typing_extensions==4.8.0
